M3L2/video-4-refactoring-a-legacy-rag-project-with-langchain-guided-walkthrough/requirements.txt
langchain>=0.1.15
openai>=0.27.0
faiss-cpu>=1.7.2
python-dotenv>=0.21.0
